+++ test-tools/proof-explorer/vite-env.d.ts
@@ -1,1 +1,2 @@
-declare const BACKEND_API: string;
+declare const VERIFIER_V0_API: string;
+declare const VERIFIER_V1_API: string;